T2884: Add ipt_NETFLOW kernel module (T75)
Add lightweight in-kernel netflow collector for operational and
security telemetry without the weight of FPC.
Implement the promisc capture patch to allow processing of data on
promiscuous interfaces for flows. This might be worthwhile
converting to a sysctl down the line, depending on use-cases (NIDS
and flow aggregators may want the opposite behavior).

Upstream at https://github.com/aabc/ipt-netflow, which was T75 way
back in may of 2016.

Testing:
  Built and used in the Arch Linux implementation. Slated for build
in Debian VyOS build container this week.